// src/api/generated/model/pollCreateRequest.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * polling API
 * polling API Document
 * OpenAPI spec version: 1.0.0
 */
import type { PollOptionCreateRequest } from './pollOptionCreateRequest';

export interface PollCreateRequest {
  pollCategory: string;
  pollDescription: string;
  pollOptions: PollOptionCreateRequest[];
  pollSeq: number;
}